Modify jobs routes such that only admin gets access to unapproved jobs
- The "/" path now shows all approved jobs, whereas "/unapproved" shows
all unapproved jobs which only the admin can access.
- Suppose some person manages to guess a particular job's id correctly, and enters
in the right url to access the job detail and do some CRUD operations.
However, the job is unapproved, and is supposed to be hidden from public view.
A 404 will be shown, to hide the fact that the job with the corresponding
id actually exists. Only admins can bypass this restriction.
checkJobIsApproved's purpose is to enforce the above.
- Submissions view calls the new "/unapproved" route.